+++ polymer/eduke32/source/jaudiolib/_midi.h
@@ -114,14 +114,14 @@ typedef struct
    short          loopcount;
    short          RunningStatus;
    unsigned       time;
-   long           FPSecondsPerTick;
+   int           FPSecondsPerTick;
    short          tick;
    short          beat;
    short          measure;
    short          BeatsPerMeasure;
    short          TicksPerBeat;
    short          TimeBase;
-   long           delay;
+   int           delay;
    short          active;
    } songcontext;
 
@@ -130,7 +130,7 @@ typedef struct
    unsigned char *start;
    unsigned char *pos;
 
-   long           delay;
+   int           delay;
    short          active;
    short          RunningStatus;
 
@@ -142,8 +142,8 @@ typedef struct
    char           EMIDI_VolumeChange;
    } track;
 
-static long _MIDI_ReadNumber( void *from, size_t size );
-static long _MIDI_ReadDelta( track *ptr );
+static int _MIDI_ReadNumber( void *from, size_t size );
+static int _MIDI_ReadDelta( track *ptr );
 static void _MIDI_ResetTracks( void );
 static void _MIDI_AdvanceTick( void );
 static void _MIDI_MetaEvent( track *Track );
