+++ polymer/eduke32/source/jaudiolib/dsoundout.c
@@ -449,7 +449,7 @@ static DWORD WINAPI isr(LPVOID parm)
             {
                 /*
                 #define copybuf(S,D,c) \
-                ({ void *__S=(S), *__D=(D); long __c=(c); \
+                ({ void *__S=(S), *__D=(D); int __c=(c); \
                 __asm__ __volatile__ ("rep; movsl" \
                 : "+S" (__S), "+D" (__D), "+c" (__c) : : "memory", "cc"); \
                 0; })
